Big reorganization
* Better name structure to indicate when things happen
* Storage trigger is lightweight and just passes to add-header CF
* add-header CF can accept some headers or a `lookup_file`, which will
look up the headers
