+++ 00-storage-fits-uploaded/requirements.txt
@@ -1,0 +1,1 @@
+requests
