+++ 01-add-header-to-db/README.md
@@ -6,6 +6,11 @@ This folder defines a [Google Cloud Function](https://cloud.google.com/functions
 This function is used to add FITS header information to the meta database 
 (which itself is a [Cloud SQL Instanct](https://cloud.google.com/sql/docs/)).
 
+This endpoint looks for two parameters, `headers` and `lookup_file`. If
+`lookup_file` is present then the header information will be pull from the file
+itself. Additionally, any `headers` will be used to update the header information
+from the file. If no `lookup_file` is found then only the `headers` will be used.
+
 > :memo: Todo: Document that explains overall structure.
 
 
@@ -13,7 +18,10 @@ Endpoint: https://us-central1-panoptes-survey.cloudfunctions.net/header-to-metad
 
 Payload: JSON message of the form: 
 	```json
-	{ 'header': dict }
+	{ 
+		'header': dict,
+		'lookup_file': str,
+	}
 	```
 
 Deploy
@@ -22,12 +30,12 @@ Deploy
 [Google Documentation](https://cloud.google.com/functions/docs/deploying/filesystem)
 
 From the directory containing the cloud function. The `entry_point` is the
-name of the function in `main.py` that we want called and `header-to-metadb`
+name of the function in `main.py` that we want called and `01-header-to-metadb`
 is the name of the Cloud Function we want to create.
 
 ```bash
 gcloud functions deploy \
-                 header-to-metadb \
+                 01-header-to-metadb \
                  --entry-point header_to_db \
                  --runtime python37 \
                  --trigger-http
