+++ 01-add-header-to-db/deploy.sh
@@ -1,7 +1,7 @@
 #!/bin/bash -e
 
 gcloud functions deploy \
-                 header-to-metadb \
+                 01-header-to-metadb \
                  --entry-point header_to_db \
                  --runtime python37 \
                  --trigger-http
