+++ 01-add-header-to-db/requirements.txt
@@ -1,0 +1,4 @@
+astropy >= 3.0.0
+google-cloud-storage
+psycopg2
+PyYAML
