--- ack-fits-received/requirements.txt
@@ -1,4 +1,0 @@
-# Function dependencies, for example:
-# package>=version
-google-cloud-storage
-requests
